cast to int

+++ gendb_english.py
@@ -197,7 +197,7 @@ class Syntactic:
 
                 raw_novel_text = self.get_raw_text(novel_id)
                 tokens = nltk.word_tokenize(raw_novel_text)
-                tokens_sum += tokens[0:self.token_size / self.num_authors]
+                tokens_sum += tokens[0:int(self.token_size / self.num_authors)]
 
                 cur.execute("INSERT INTO section VALUES(%s,%s,%s,%s,%s)", [i + 1, num_section,
                                                                            raw_novel_text, novel_id,
